% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byRegion.r
\name{byRegion}
\alias{byRegion}
\title{Pokemon Within Region}
\usage{
byRegion(region = NA, dexNum = NA)
}
\arguments{
\item{region}{String indicating a pokemon region. See regions() for more list}

\item{dexNum}{integer}
}
\value{

}
\description{
Given a region and region-specific pokedex number, the function returns the associated pokemon. If either parameters are not specified, random ones are used
}
\examples{
byRegion()
byRegion(region = "kanto")
byRegion(region = "national", 3)

}
