% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectiveness.r
\name{effectiveness}
\alias{effectiveness}
\title{Type Effectiveness}
\usage{
effectiveness(attack, defend)
}
\arguments{
\item{attack}{String. The type of the attacking (your own) pokemon}

\item{defend}{String. The type of the defending (opponent's) pokemon}
}
\value{
Text describing the effectiveness. Maybe better as numeric?
}
\description{
Describes the damage effectiveness of two Pokemon Types. For a list of types, see pokeTypes()
}
\examples{
effectiveness("water", "fire")
effectiveness("fire", "water")
}
