% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolution.r
\name{evolution}
\alias{evolution}
\title{Next Pokemon Evolution}
\usage{
evolution(id)
}
\arguments{
\item{id}{integer. The id of the requested pokemon}
}
\value{
String name of the entered pokemon's evolution
}
\description{
Returns the next stage of evolution for the entered pokemon ID
}
\examples{
evolution(1)
}
