% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaviest.r
\name{heaviest}
\alias{heaviest}
\title{Team by weight}
\usage{
heaviest(myTeam = rTeam())
}
\arguments{
\item{myTeam}{A list of string names of pokemon}
}
\value{
Dataframe of pokemon and corresponding weight
}
\description{
Given a list of pokemon names (a "team"), the function will output the team sorted by weight. If no list is specified, the function will generate a random team of five.
}
\examples{
heaviest()
heaviest(c("piakachu", "lotad"))
}
