% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokehack.R
\name{pokehack}
\alias{pokehack}
\title{Boost Pokémon Stats}
\usage{
pokehack(pokemon_name, data)
}
\arguments{
\item{pokemon_name}{Character. Name of the Pokémon to boost.}

\item{data}{A data frame containing Pokémon data with columns:
"name", "defense", "sp_attack", "sp_defense", and "speed".}
}
\value{
A modified data frame with the updated stats for the specified Pokémon.
}
\description{
This function takes a Pokémon's name from the "name" column of a data frame
and increases the specified attributes (defense, sp_attack, sp_defense, speed)
by 10\%.
}
\examples{
pokedata <- pokehack("Bulbasaur", pokedata)
}
