% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poketrain.R
\name{poketrain}
\alias{poketrain}
\title{Train Pokémon for a Battle}
\usage{
poketrain(selected_pokemon, data)
}
\arguments{
\item{selected_pokemon}{Character vector. Names of the 6 Pokémon to train.}

\item{data}{A data frame containing Pokémon data with at least the following columns:
"name", "weight_kg", and "experience_growth".}
}
\value{
A modified data frame with updated \code{weight_kg} and \code{experience_growth} for the selected Pokémon.
}
\description{
This function allows a user to select 6 Pokémon for training before a battle.
The selected Pokémon lose 10\% of their body weight (\code{weight_kg}) and gain 5\% experience (\code{experience_growth}).
}
\examples{
trained_pokedata <- poketrain(
  selected_pokemon = c("Bulbasaur", "Charmander", "Squirtle", "Pikachu", "Jigglypuff", "Meowth"),
  data = pokedata
)
}
