% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokeTypes.r
\name{pokeTypes}
\alias{pokeTypes}
\title{Pokemon Types}
\usage{
pokeTypes()
}
\value{
List of strings containing all pokemon types
}
\description{
A helper function to display all available Pokemon Types. It's a one liner so I didn't really consider it part of the five.
}
