% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.r
\name{regions}
\alias{regions}
\title{Pokemon Game Regions}
\usage{
regions()
}
\value{
A list of Strings containing all pokemon regions available
}
\description{
A helper function to display all available Pokemon Regions. Again, it's a one liner so I didn't really consider it part of the five.
}
