% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTeam.r
\name{rTeam}
\alias{rTeam}
\title{Random Team}
\usage{
rTeam(n = 5)
}
\arguments{
\item{n}{an integer indicating how big your team will be. Must be less than 899}
}
\value{
A list of strings
}
\description{
Generates a random team of pokemon
}
\examples{
rTeam()
rTeam(2)


}
